+++ src/main/java/com/sparta/springjpascheduler/service/TodoService.java
@@ -5,23 +5,22 @@ import com.sparta.springjpascheduler.dto.TodoResponseDto;
 import com.sparta.springjpascheduler.entity.Todo;
 import com.sparta.springjpascheduler.repository.CommentRepository;
 import com.sparta.springjpascheduler.repository.TodoRepository;
+import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
-import org.springframework.transaction.annotation.Transactional;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Optional;
 
 @Service
+@RequiredArgsConstructor
 public class TodoService {
 
-    @Autowired
-    private TodoRepository todoRepository;
-    @Autowired
-    private CommentRepository commentRepository;
+    private final TodoRepository todoRepository;
+    private final CommentRepository commentRepository;
 
     @Transactional
     public TodoResponseDto saveTodo(TodoRequestDto todoRequestDto) {
@@ -35,6 +34,10 @@ public class TodoService {
 
     @Transactional(readOnly = true)
     public Optional<TodoResponseDto> getTodoById(Long id) {
+        if (!todoRepository.existsById(id)) {
+            throw new RuntimeException("Todo not found");
+        }
+
         return todoRepository.findById(id).map(this::mapToResponseDto);
     }
 
@@ -50,30 +53,33 @@ public class TodoService {
         return mapToResponseDto(updatedTodo);
     }
 
+    @Transactional(readOnly = true)
     public Page<TodoResponseDto> getTodos(int page, int size) {
-        Sort sort = Sort.by(Sort.Direction.DESC, "updatedAt");
+        Sort sort = Sort.by(Sort.Direction.DESC, "modifiedAt");
         Pageable pageable = PageRequest.of(page, size, sort);
 
         Page<Todo> todoList = todoRepository.findAll(pageable);
         return todoList.map(this::mapToResponseDto);
     }
 
+    @Transactional
     public void deleteTodo(Long id) {
+        if (!todoRepository.existsById(id)) {
+            throw new RuntimeException("Todo not found");
+        }
         todoRepository.deleteById(id);
     }
 
     private TodoResponseDto mapToResponseDto(Todo todo) {
-        TodoResponseDto dto = new TodoResponseDto();
-        dto.setId(todo.getId());
-        dto.setUserName(todo.getUserName());
-        dto.setTitle(todo.getTitle());
-        dto.setContent(todo.getContent());
-        dto.setCreatedAt(todo.getCreatedAt());
-        dto.setUpdatedAt(todo.getUpdatedAt());
 
-        int commentCnt = commentRepository.countByTodoId(todo.getId());
-        dto.setCommentCnt(commentCnt);
-
-        return dto;
+        return new TodoResponseDto(
+                todo.getId(),
+                todo.getUserName(),
+                todo.getTitle(),
+                todo.getContent(),
+                todo.getCreatedAt(),
+                todo.getModifiedAt(),
+                commentRepository.countByTodoId(todo.getId())
+        );
     }
 }
